+++ mod-source-record-manager-server/src/main/java/org/folio/services/mappers/processor/parameters/MappingParameters.java
@@ -1,6 +1,7 @@
 package org.folio.services.mappers.processor.parameters;
 
 import org.apache.commons.collections4.list.UnmodifiableList;
+import org.folio.rest.jaxrs.model.AlternativeTitleType;
 import org.folio.rest.jaxrs.model.ClassificationType;
 import org.folio.rest.jaxrs.model.ContributorNameType;
 import org.folio.rest.jaxrs.model.ElectronicAccessRelationship;
@@ -26,6 +27,7 @@ public class MappingParameters {
   private UnmodifiableList<ContributorType> contributorTypes;
   private UnmodifiableList<ContributorNameType> contributorNameTypes;
   private UnmodifiableList<InstanceNoteType> instanceNoteTypes;
+  private UnmodifiableList<AlternativeTitleType> alternativeTitleTypes;
 
   public boolean isInitialized() {
     return initializedState;
@@ -105,6 +107,15 @@ public class MappingParameters {
 
   public MappingParameters withInstanceNoteTypes(List<InstanceNoteType> instanceNoteTypes) {
     this.instanceNoteTypes = new UnmodifiableList<>(instanceNoteTypes);
+    return this;
+  }
+
+  public List<AlternativeTitleType> getAlternativeTitleTypes() {
+    return alternativeTitleTypes;
+  }
+
+  public MappingParameters withAlternativeTitleTypes(List<AlternativeTitleType> alternativeTitleTypes) {
+    this.alternativeTitleTypes = new UnmodifiableList<>(alternativeTitleTypes);
     return this;
   }
 }